refactor(proxy): unify response inspection and delivery lifecycle
