+++ worker/proxy.ts
@@ -16,7 +16,8 @@ import {
 } from "./providers";
 import { applyTransportHeaders, transformTransportBody } from "./provider-auth.ts";
 import { actualModelCost, estimateModelCost } from "./pricing";
-import { extractSseUsageTokens, extractUsageTokens, type UsageTokens } from "./token-usage";
+import type { UsageTokens } from "./token-usage";
+import { normalizePreStreamError, observeUsage } from "./proxy-response";
 import type { AuthorizedIdentity, CompiledEndpoint, CompiledModel, CompiledProvider, CompiledQuotaConfig, Env, ProviderConnection, UsageEvent } from "./types";
 import {
   errorResponse, HttpError, parseProxyKey, randomId, readJson, safeEqual, sha256Hex,
@@ -387,18 +388,9 @@ async function proxySelected(request: Request, env: Env, context: ExecutionConte
     return errorResponse("provider_unavailable", `upstream request to provider ${selection.provider.id} failed`, 502, undefined);
   }
   clearTimeout(timeout);
-  const contentType = response.headers.get("content-type") ?? "";
-  if (contentType.includes("json") || contentType.includes("text/event-stream")) {
-    context.waitUntil(finalizeResponse(response.clone(), env, auth, selection, request, requestId, started, cost, reservation, content, compound));
-  } else {
-    // Draining a tee for accounting buffers the entire binary transfer behind a
-    // slow client. Observe its completion without creating a second consumer.
-    const upstreamResponse = response;
-    let completed!: () => void;
-    const completion = new Promise<void>(resolve => { completed = resolve; });
-    response = observeResponseBody(response, completed);
-    context.waitUntil(completion.then(() => finalizeResponse(upstreamResponse, env, auth, selection, request, requestId, started, cost, reservation, content, compound)));
-  }
+  const observed = observeUsage(response);
+  context.waitUntil(observed.tokens.then(tokens => finalizeResponse(response, tokens, env, auth, selection, request, requestId, started, cost, reservation, content, compound)));
+  response = observed.response;
   const outputHeaders = new Headers(response.headers);
   for (const name of ["connection", "keep-alive", "proxy-authenticate", "proxy-authorization", "set-cookie", "trailer", "transfer-encoding", "upgrade"]) outputHeaders.delete(name);
   outputHeaders.set("x-clawrouter-upstream-provider", selection.provider.id);
@@ -406,163 +398,6 @@ async function proxySelected(request: Request, env: Env, context: ExecutionConte
   if (grantFailover) outputHeaders.set("x-clawrouter-grant-failover", "1");
   outputHeaders.set("x-clawrouter-content-retention", auth.policy.retainRequestContent !== false && !auth.contentRetentionDisabled ? "on; retention-days=30" : "off");
   return new Response(response.body, { status: response.status, statusText: response.statusText, headers: outputHeaders });
-}
-
-export async function normalizePreStreamError(response: Response, streamingRequested: boolean): Promise<Response> {
-  if (!streamingRequested) return response;
-  const eventStream = response.headers.get("content-type")?.toLowerCase().includes("text/event-stream") === true;
-  if (response.status >= 400) {
-    if (eventStream && response.body) return normalizeFirstSseEvent(response, response.status);
-    const body = await readLimited(response, 64 * 1024).catch(() => "");
-    return mappedUpstreamError(response, upstreamError(body), response.status);
-  }
-  if (!response.ok || !eventStream || !response.body) return response;
-  return normalizeFirstSseEvent(response, null);
-}
-
-const FIRST_SSE_EVENT_LIMIT = 8 * 1024;
-
-async function normalizeFirstSseEvent(response: Response, errorStatus: number | null): Promise<Response> {
-  const reader = response.body!.getReader();
-  const chunks: Uint8Array[] = [];
-  const sniffed = new Uint8Array(FIRST_SSE_EVENT_LIMIT);
-  let sniffedLength = 0;
-  let eventStart = 0;
-  while (true) {
-    const { done, value } = await reader.read();
-    if (value?.byteLength) {
-      chunks.push(value);
-      const copyLength = Math.min(value.byteLength, FIRST_SSE_EVENT_LIMIT - sniffedLength);
-      if (copyLength > 0) {
-        sniffed.set(value.subarray(0, copyLength), sniffedLength);
-        sniffedLength += copyLength;
-      }
-    }
-    while (eventStart < sniffedLength) {
-      const boundary = sseEventBoundary(sniffed, eventStart, sniffedLength);
-      if (!boundary) break;
-      const event = classifySseEvent(sniffed.subarray(eventStart, boundary.start));
-      eventStart = boundary.end;
-      if (event.kind === "empty") continue;
-      if (errorStatus !== null || event.kind === "error") {
-        await reader.cancel().catch(() => undefined);
-        const upstream = event.upstream;
-        const status = errorStatus ?? (typeof upstream.code === "number" && Number.isInteger(upstream.code) && upstream.code >= 400 && upstream.code <= 599 ? upstream.code : 502);
-        return mappedUpstreamError(response, upstream, status);
-      }
-      return replayResponse(response, reader, chunks, done);
-    }
-    if (done || sniffedLength === FIRST_SSE_EVENT_LIMIT) {
-      if (errorStatus === null) return replayResponse(response, reader, chunks, done);
-      if (!done) await reader.cancel().catch(() => undefined);
-      return mappedUpstreamError(response, {}, errorStatus);
-    }
-  }
-}
-
-function replayResponse(response: Response, reader: ReadableStreamDefaultReader<Uint8Array>, chunks: Uint8Array[], readerDone: boolean): Response {
-  let chunkIndex = 0;
-  const body = new ReadableStream<Uint8Array>({
-    async pull(controller) {
-      if (chunkIndex < chunks.length) {
-        controller.enqueue(chunks[chunkIndex++]);
-        return;
-      }
-      if (readerDone) {
-        controller.close();
-        return;
-      }
-      try {
-        const { done, value } = await reader.read();
-        if (done) controller.close();
-        else controller.enqueue(value);
-      } catch (error) { controller.error(error); }
-    },
-    cancel(reason) { return reader.cancel(reason); },
-  });
-  return new Response(body, { status: response.status, statusText: response.statusText, headers: response.headers });
-}
-
-function sseEventBoundary(bytes: Uint8Array, start: number, length: number): { start: number; end: number } | null {
-  let lineStart = start;
-  for (let index = start; index < length; index += 1) {
-    if (bytes[index] !== 10 && bytes[index] !== 13) continue;
-    const next = bytes[index] === 13 && index + 1 < length && bytes[index + 1] === 10 ? index + 2 : index + 1;
-    if (index === lineStart) return { start: lineStart, end: next };
-    lineStart = next;
-    index = next - 1;
-  }
-  return null;
-}
-
-function classifySseEvent(bytes: Uint8Array): { kind: "empty" } | { kind: "healthy" | "error"; upstream: ReturnType<typeof upstreamError> } {
-  let eventType = "";
-  const dataLines: string[] = [];
-  for (const line of new TextDecoder().decode(bytes).split(/\r\n|\r|\n/)) {
-    if (!line || line.startsWith(":")) continue;
-    const colon = line.indexOf(":");
-    const field = colon < 0 ? line : line.slice(0, colon);
-    let value = colon < 0 ? "" : line.slice(colon + 1);
-    if (value.startsWith(" ")) value = value.slice(1);
-    if (field === "event") eventType = value;
-    else if (field === "data") dataLines.push(value);
-  }
-  const data = dataLines.join("\n");
-  if (dataLines.length === 0) return { kind: "empty" };
-  const upstream = upstreamError(data);
-  return eventType === "error" || hasTopLevelError(data) ? { kind: "error", upstream } : { kind: "healthy", upstream };
-}
-
-function hasTopLevelError(data: string): boolean {
-  try {
-    const value: unknown = JSON.parse(data);
-    return !!value && typeof value === "object" && !Array.isArray(value) && Object.prototype.hasOwnProperty.call(value, "error");
-  } catch { return false; }
-}
-
-function mappedUpstreamError(response: Response, upstream: ReturnType<typeof upstreamError>, status: number): Response {
-  const headers = new Headers(response.headers);
-  headers.delete("content-length");
-  headers.set("content-type", "application/json; charset=utf-8");
-  return Response.json({ error: {
-    message: upstream.message ?? (response.ok ? "upstream request failed" : response.statusText || "upstream request failed"),
-    type: upstream.type ?? "upstream_error",
-    code: upstream.code ?? status,
-  } }, { status, statusText: status === response.status ? response.statusText : "", headers });
-}
-
-function upstreamError(body: string): { message?: string; type?: string; code?: string | number } {
-  let value: unknown;
-  try {
-    const eventData = firstSseData(body);
-    value = JSON.parse(eventData || body);
-  } catch { return {}; }
-  if (!value || typeof value !== "object" || Array.isArray(value)) return {};
-  const error = "error" in value && value.error && typeof value.error === "object" && !Array.isArray(value.error) ? value.error as Record<string, unknown> : value as Record<string, unknown>;
-  return {
-    message: typeof error.message === "string" ? error.message : undefined,
-    type: typeof error.type === "string" ? error.type : undefined,
-    code: typeof error.code === "string" || typeof error.code === "number" ? error.code : undefined,
-  };
-}
-
-function firstSseData(body: string): string | null {
-  const dataLines: string[] = [];
-  for (const line of body.split(/\r\n|\r|\n/)) {
-    if (!line) {
-      if (dataLines.some((value) => value !== "")) break;
-      dataLines.length = 0;
-      continue;
-    }
-    if (line.startsWith(":")) continue;
-    const colon = line.indexOf(":");
-    const field = colon < 0 ? line : line.slice(0, colon);
-    if (field !== "data") continue;
-    let value = colon < 0 ? "" : line.slice(colon + 1);
-    if (value.startsWith(" ")) value = value.slice(1);
-    dataLines.push(value);
-  }
-  return dataLines.length > 0 ? dataLines.join("\n") : null;
 }
 
 async function reserveSelected(request: Request, env: Env, context: ExecutionContext, mode: AuthMode, selection: ProxySelection, preauthenticated: AuthorizedIdentity | null, compound?: CompoundRequestContext): Promise<ReservedProxyBudget | Response> {
@@ -657,13 +492,7 @@ async function preauthenticate(request: Request, env: Env, mode: AuthMode, provi
   return providerId ? accessIdentity(request, env, providerId) : null;
 }
 
-async function finalizeResponse(response: Response, env: Env, auth: AuthorizedIdentity, selection: ProxySelection, request: Request, requestId: string, started: number, estimated: Cost, reservation: BudgetReservation, content: string | null, compound?: CompoundRequestContext): Promise<void> {
-  let tokens: Tokens | null = null;
-  try {
-    const contentType = response.headers.get("content-type") ?? "";
-    if (contentType.includes("json")) tokens = extractUsageTokens(JSON.parse(await readLimited(response, 2 * 1024 * 1024)));
-    else if (contentType.includes("text/event-stream")) tokens = extractSseUsageTokens(await readLimited(response, 2 * 1024 * 1024));
-  } catch { /* usage is best-effort; reservation stays conservative */ }
+async function finalizeResponse(response: Response, tokens: Tokens | null, env: Env, auth: AuthorizedIdentity, selection: ProxySelection, request: Request, requestId: string, started: number, estimated: Cost, reservation: BudgetReservation, content: string | null, compound?: CompoundRequestContext): Promise<void> {
   const measured = tokens ? actualCost(selection.model, tokens, auth.policy.requestCostMicros) : null;
   const actual = response.ok && measured != null ? measured : response.ok ? estimated.reserveMicros : 0;
   await finalizeAccounting(env, auth, reservation, actual, usageEvent(auth, selection, request, requestId, started, response.status, tokens, estimated, reservation, content, response.ok ? "success" : response.status < 500 ? "client_error" : "provider_error", actual, compound));
@@ -707,40 +536,6 @@ function usageEvent(auth: AuthorizedIdentity, selection: ProxySelection, request
     pricing_effective_at: selection.model?.pricing?.effectiveAt ?? null, cost_basis: cost.basis, status_code: statusCode,
     duration_ms: Date.now() - started, content_retained: !!contentRef, content_ref: contentRef, status,
   };
-}
-
-async function readLimited(response: Response, limit: number): Promise<string> {
-  if (!response.body) return "";
-  const reader = response.body.getReader(), decoder = new TextDecoder();
-  let size = 0, text = "";
-  try {
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      size += value.byteLength;
-      if (size > limit) throw new Error("usage payload exceeds inspection limit");
-      text += decoder.decode(value, { stream: true });
-    }
-    return text + decoder.decode();
-  } finally { if (size > limit) await reader.cancel(); }
-}
-
-function observeResponseBody(response: Response, complete: () => void): Response {
-  if (!response.body) { complete(); return response; }
-  const reader = response.body.getReader();
-  let finished = false;
-  const finish = () => { if (!finished) { finished = true; reader.releaseLock(); complete(); } };
-  const body = new ReadableStream<Uint8Array>({
-    async pull(controller) {
-      try {
-        const next = await reader.read();
-        if (next.done) { finish(); controller.close(); }
-        else controller.enqueue(next.value);
-      } catch (error) { finish(); controller.error(error); }
-    },
-    async cancel(reason) { try { await reader.cancel(reason); } finally { finish(); } },
-  }, { highWaterMark: 0 });
-  return new Response(body, { status: response.status, statusText: response.statusText, headers: response.headers });
 }
 
 function searchParamsRecord(params: URLSearchParams): Record<string, string> { const result: Record<string, string> = {}; params.forEach((value, key) => { result[key] = value; }); return result; }
